+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - If you insert a footnote, the cursor is not moved throughout the document so that the writing flow is more immersive.
 - The text field used to edit a footnote reference text is now automatically focused.
 - The editor will now directly mute lines when in full screen as soon as you change the preference setting for this. You don't have to move the cursor anymore for this.
+- Fixed a bug that showed a dedicated _file_ menu when right-clicking directory ribbons and then threw errors if you clicked on anything in the menu.
+- Fixed the strange behaviour of Zettlr telling you there are no suggestions for a word, although you did not right-click a wrongly spelled word.
 
 ## Under the hood
 
@@ -14,6 +16,7 @@
 - Changed the event type the editor is listening to when you finished editing a footnote from `keyup` to `keydown`.
 - Moved the inline and block element rendering functions to their own commands to reduce the file size of `ZettlrEditor`.
 - Fixed the task recognition regex to prevent ugly errors logged to the console if you entered on an empty line a task list item directly followed by braces (such as `- [ ](some text)`).
+- Additional security checks while building the context menu.
 
 # 0.18.3