+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.18.4
+# 0.19.0
 
 ## GUI and functionality